requirements.txt
requests~=2.27.1
python-dotenv~=0.19.2
retry~=0.9.2
urllib3~=1.26.8
pymongo~=4.0.2
matplotlib~=3.5.1
networkx~=2.7.1
scipy~=1.8.0